seperate jsb_adpf register function
